Refactor diagnostic

+++ projet/lib/include/defines.h
@@ -28,7 +28,7 @@ const uint8_t VERT = 0x01;
 // Informations sur l'equipe
 const char INFO_NOM_ROBOT[] = "Thinky Winky";   ///< le nom du robot
 const char INFO_EQUIPE[] = "68116";             ///< le numero de l'equipe
-const int INFO_GROUPE = 3;                      ///< le numero du groupe
+const uint8_t INFO_GROUPE = 3;                  ///< le numero du groupe
 const char INFO_SESSION[] = "18-1";             ///< l'annee-session
 const uint8_t COULEUR_ROBOT1 = 1;               ///< la couleur du premier robot
 const uint8_t COULEUR_ROBOT2 = 4;               ///< la couleur du deuxieme robot
